extract posix functions from std/os.zig to std/os/posix.zig
See #2380
