+++ CMakeLists.txt
@@ -621,9 +621,9 @@ set(ZIG_STD_FILES
     "os/time.zig"
     "os/uefi.zig"
     "os/wasi.zig"
-    "os/wasi/core.zig"
     "os/windows.zig"
     "os/windows/advapi32.zig"
+    "os/windows/errno.zig"
     "os/windows/error.zig"
     "os/windows/kernel32.zig"
     "os/windows/ntdll.zig"
