+++ example/hello_world/hello.zig
@@ -1,9 +1,5 @@
 const std = @import("std");
 
-pub fn main() !void {
-    // If this program is run without stdout attached, exit with an error.
-    const stdout_file = try std.io.getStdOut();
-    // If this program encounters pipe failure when printing to stdout, exit
-    // with an error.
-    try stdout_file.write("Hello, world!\n");
+pub fn main() void {
+    std.debug.warn("Hello, world!\n");
 }
