+++ std/os.zig
@@ -2,10 +2,6 @@ const std = @import("std.zig");
 const builtin = @import("builtin");
 const Os = builtin.Os;
 const is_windows = builtin.os == Os.windows;
-const is_posix = switch (builtin.os) {
-    builtin.Os.linux, builtin.Os.macosx, builtin.Os.freebsd, builtin.Os.netbsd => true,
-    else => false,
-};
 const os = @This();
 
 comptime {
@@ -36,13 +32,13 @@ pub const zen = @import("os/zen.zig");
 pub const uefi = @import("os/uefi.zig");
 pub const wasi = @import("os/wasi.zig");
 
-pub const posix = switch (builtin.os) {
-    Os.linux => linux,
-    Os.macosx, Os.ios => darwin,
-    Os.freebsd => freebsd,
-    Os.netbsd => netbsd,
-    Os.zen => zen,
-    Os.wasi => wasi,
+pub const system = if (builtin.link_libc) c else switch (builtin.os) {
+    .linux => linux,
+    .macosx, .ios, .watchos, .tvos => darwin,
+    .freebsd => freebsd,
+    .netbsd => netbsd,
+    .zen => zen,
+    .wasi => wasi,
     else => @compileError("Unsupported OS"),
 };
 
@@ -58,13 +54,17 @@ pub const page_size = switch (builtin.arch) {
     else => 4 * 1024,
 };
 
+/// This represents the maximum size of a UTF-8 encoded file path.
+/// All file system operations which return a path are guaranteed to
+/// fit into a UTF-8 encoded array of this length.
+/// path being too long if it is this 0long
 pub const MAX_PATH_BYTES = switch (builtin.os) {
-    Os.linux, Os.macosx, Os.ios, Os.freebsd, Os.netbsd => posix.PATH_MAX,
+    .linux, .macosx, .ios, .freebsd, .netbsd => posix.PATH_MAX,
     // Each UTF-16LE character may be expanded to 3 UTF-8 bytes.
     // If it would require 4 UTF-8 bytes, then there would be a surrogate
     // pair in the UTF-16LE, and we (over)account 3 bytes for it that way.
     // +1 for the null byte at the end, which can be encoded in 1 byte.
-    Os.windows => windows_util.PATH_MAX_WIDE * 3 + 1,
+    .windows => posix.PATH_MAX_WIDE * 3 + 1,
     else => @compileError("Unsupported OS"),
 };
 
@@ -98,6 +98,22 @@ pub const FileHandle = if (is_windows) windows.HANDLE else i32;
 pub const getAppDataDir = @import("os/get_app_data_dir.zig").getAppDataDir;
 pub const GetAppDataDirError = @import("os/get_app_data_dir.zig").GetAppDataDirError;
 
+pub const getRandomBytes = posix.getrandom;
+pub const abort = posix.abort;
+pub const exit = posix.exit;
+pub const symLink = posix.symlink;
+pub const symLinkC = posix.symlinkC;
+pub const symLinkW = posix.symlinkW;
+pub const deleteFile = posix.unlink;
+pub const deleteFileC = posix.unlinkC;
+pub const deleteFileW = posix.unlinkW;
+pub const rename = posix.rename;
+pub const renameC = posix.renameC;
+pub const renameW = posix.renameW;
+pub const changeCurDir = posix.chdir;
+pub const changeCurDirC = posix.chdirC;
+pub const changeCurDirW = posix.chdirW;
+
 const debug = std.debug;
 const assert = debug.assert;
 const testing = std.testing;
@@ -115,610 +131,6 @@ const base64 = std.base64;
 const ArrayList = std.ArrayList;
 const Buffer = std.Buffer;
 const math = std.math;
-
-/// Fills `buf` with random bytes. If linking against libc, this calls the
-/// appropriate OS-specific library call. Otherwise it uses the zig standard
-/// library implementation.
-pub fn getRandomBytes(buf: []u8) !void {
-    switch (builtin.os) {
-        Os.linux => while (true) {
-            // TODO check libc version and potentially call c.getrandom.
-            // See #397
-            const errno = posix.getErrno(posix.getrandom(buf.ptr, buf.len, 0));
-            switch (errno) {
-                0 => return,
-                posix.EINVAL => unreachable,
-                posix.EFAULT => unreachable,
-                posix.EINTR => continue,
-                posix.ENOSYS => return getRandomBytesDevURandom(buf),
-                else => return unexpectedErrorPosix(errno),
-            }
-        },
-        Os.macosx, Os.ios, Os.freebsd, Os.netbsd => return getRandomBytesDevURandom(buf),
-        Os.windows => {
-            // Call RtlGenRandom() instead of CryptGetRandom() on Windows
-            // https://github.com/rust-lang-nursery/rand/issues/111
-            // https://bugzilla.mozilla.org/show_bug.cgi?id=504270
-            if (windows.RtlGenRandom(buf.ptr, buf.len) == 0) {
-                const err = windows.GetLastError();
-                return switch (err) {
-                    else => unexpectedErrorWindows(err),
-                };
-            }
-        },
-        Os.wasi => {
-            const random_get_result = os.wasi.random_get(buf.ptr, buf.len);
-            if (random_get_result != os.wasi.ESUCCESS) {
-                return error.Unknown;
-            }
-        },
-        Os.zen => {
-            const randomness = []u8{ 42, 1, 7, 12, 22, 17, 99, 16, 26, 87, 41, 45 };
-            var i: usize = 0;
-            while (i < buf.len) : (i += 1) {
-                if (i > randomness.len) return error.Unknown;
-                buf[i] = randomness[i];
-            }
-        },
-        else => @compileError("Unsupported OS"),
-    }
-}
-
-fn getRandomBytesDevURandom(buf: []u8) !void {
-    const fd = try posixOpenC(c"/dev/urandom", posix.O_RDONLY | posix.O_CLOEXEC, 0);
-    defer close(fd);
-
-    const stream = &File.openHandle(fd).inStream().stream;
-    stream.readNoEof(buf) catch |err| switch (err) {
-        error.EndOfStream => unreachable,
-        error.OperationAborted => unreachable,
-        error.BrokenPipe => unreachable,
-        error.Unexpected => return error.Unexpected,
-        error.InputOutput => return error.Unexpected,
-        error.SystemResources => return error.Unexpected,
-        error.IsDir => unreachable,
-    };
-}
-
-test "os.getRandomBytes" {
-    var buf_a: [50]u8 = undefined;
-    var buf_b: [50]u8 = undefined;
-    // Call Twice
-    try getRandomBytes(buf_a[0..]);
-    try getRandomBytes(buf_b[0..]);
-
-    // Check if random (not 100% conclusive)
-    testing.expect(!mem.eql(u8, buf_a, buf_b));
-}
-
-/// Raises a signal in the current kernel thread, ending its execution.
-/// If linking against libc, this calls the abort() libc function. Otherwise
-/// it uses the zig standard library implementation.
-pub fn abort() noreturn {
-    @setCold(true);
-    if (builtin.link_libc) {
-        c.abort();
-    }
-    switch (builtin.os) {
-        Os.linux, Os.macosx, Os.ios, Os.freebsd, Os.netbsd => {
-            _ = posix.raise(posix.SIGABRT);
-            _ = posix.raise(posix.SIGKILL);
-            while (true) {}
-        },
-        Os.windows => {
-            if (builtin.mode == builtin.Mode.Debug) {
-                @breakpoint();
-            }
-            windows.ExitProcess(3);
-        },
-        Os.wasi => {
-            _ = wasi.proc_raise(wasi.SIGABRT);
-            // TODO: Is SIGKILL even necessary?
-            _ = wasi.proc_raise(wasi.SIGKILL);
-            while (true) {}
-        },
-        Os.uefi => {
-            // TODO there's gotta be a better thing to do here than loop forever
-            while (true) {}
-        },
-        else => @compileError("Unsupported OS"),
-    }
-}
-
-/// Exits the program cleanly with the specified status code.
-pub fn exit(status: u8) noreturn {
-    @setCold(true);
-    if (builtin.link_libc) {
-        c.exit(status);
-    }
-    switch (builtin.os) {
-        Os.linux => {
-            if (builtin.single_threaded) {
-                linux.exit(status);
-            } else {
-                linux.exit_group(status);
-            }
-        },
-        Os.macosx, Os.ios, Os.freebsd, Os.netbsd => {
-            posix.exit(status);
-        },
-        Os.windows => {
-            windows.ExitProcess(status);
-        },
-        Os.wasi => {
-            wasi.proc_exit(status);
-        },
-        else => @compileError("Unsupported OS"),
-    }
-}
-
-/// When a file descriptor is closed on linux, it pops the first
-/// node from this queue and resumes it.
-/// Async functions which get the EMFILE error code can suspend,
-/// putting their coroutine handle into this list.
-/// TODO make this an atomic linked list
-pub var emfile_promise_queue = std.LinkedList(promise).init();
-
-/// Closes the file handle. Keeps trying if it gets interrupted by a signal.
-pub fn close(handle: FileHandle) void {
-    if (is_windows) {
-        windows_util.windowsClose(handle);
-    } else {
-        while (true) {
-            const err = posix.getErrno(posix.close(handle));
-            switch (err) {
-                posix.EINTR => continue,
-                else => {
-                    if (emfile_promise_queue.popFirst()) |p| resume p.data;
-                    return;
-                },
-            }
-        }
-    }
-}
-
-pub const PosixReadError = error{
-    InputOutput,
-    SystemResources,
-    IsDir,
-    Unexpected,
-};
-
-/// Returns the number of bytes that were read, which can be less than
-/// buf.len. If 0 bytes were read, that means EOF.
-pub fn posixRead(fd: i32, buf: []u8) PosixReadError!usize {
-    // Linux can return EINVAL when read amount is > 0x7ffff000
-    // See https://github.com/ziglang/zig/pull/743#issuecomment-363158274
-    const max_buf_len = 0x7ffff000;
-
-    var index: usize = 0;
-    while (index < buf.len) {
-        const want_to_read = math.min(buf.len - index, usize(max_buf_len));
-        const rc = posix.read(fd, buf.ptr + index, want_to_read);
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0 => {
-                index += rc;
-                if (rc == want_to_read) continue;
-                // Read returned less than buf.len.
-                return index;
-            },
-            posix.EINTR => continue,
-            posix.EINVAL => unreachable,
-            posix.EFAULT => unreachable,
-            posix.EAGAIN => unreachable,
-            posix.EBADF => unreachable, // always a race condition
-            posix.EIO => return error.InputOutput,
-            posix.EISDIR => return error.IsDir,
-            posix.ENOBUFS => return error.SystemResources,
-            posix.ENOMEM => return error.SystemResources,
-            else => return unexpectedErrorPosix(err),
-        }
-    }
-    return index;
-}
-
-/// Number of bytes read is returned. Upon reading end-of-file, zero is returned.
-pub fn posix_preadv(fd: i32, iov: [*]const posix.iovec, count: usize, offset: u64) PosixReadError!usize {
-    switch (builtin.os) {
-        builtin.Os.macosx => {
-            // Darwin does not have preadv but it does have pread.
-            var off: usize = 0;
-            var iov_i: usize = 0;
-            var inner_off: usize = 0;
-            while (true) {
-                const v = iov[iov_i];
-                const rc = darwin.pread(fd, v.iov_base + inner_off, v.iov_len - inner_off, offset + off);
-                const err = darwin.getErrno(rc);
-                switch (err) {
-                    0 => {
-                        off += rc;
-                        inner_off += rc;
-                        if (inner_off == v.iov_len) {
-                            iov_i += 1;
-                            inner_off = 0;
-                            if (iov_i == count) {
-                                return off;
-                            }
-                        }
-                        if (rc == 0) return off; // EOF
-                        continue;
-                    },
-                    posix.EINTR => continue,
-                    posix.EINVAL => unreachable,
-                    posix.EFAULT => unreachable,
-                    posix.ESPIPE => unreachable, // fd is not seekable
-                    posix.EAGAIN => unreachable, // this function is not for non blocking
-                    posix.EBADF => unreachable, // always a race condition
-                    posix.EIO => return error.InputOutput,
-                    posix.EISDIR => return error.IsDir,
-                    posix.ENOBUFS => return error.SystemResources,
-                    posix.ENOMEM => return error.SystemResources,
-                    else => return unexpectedErrorPosix(err),
-                }
-            }
-        },
-        builtin.Os.linux, builtin.Os.freebsd, Os.netbsd => while (true) {
-            const rc = posix.preadv(fd, iov, count, offset);
-            const err = posix.getErrno(rc);
-            switch (err) {
-                0 => return rc,
-                posix.EINTR => continue,
-                posix.EINVAL => unreachable,
-                posix.EFAULT => unreachable,
-                posix.EAGAIN => unreachable, // don't call this function for non blocking
-                posix.EBADF => unreachable, // always a race condition
-                posix.EIO => return error.InputOutput,
-                posix.EISDIR => return error.IsDir,
-                posix.ENOBUFS => return error.SystemResources,
-                posix.ENOMEM => return error.SystemResources,
-                else => return unexpectedErrorPosix(err),
-            }
-        },
-        else => @compileError("Unsupported OS"),
-    }
-}
-
-pub const PosixWriteError = error{
-    DiskQuota,
-    FileTooBig,
-    InputOutput,
-    NoSpaceLeft,
-    AccessDenied,
-    BrokenPipe,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-/// Calls POSIX write, and keeps trying if it gets interrupted.
-pub fn posixWrite(fd: i32, bytes: []const u8) PosixWriteError!void {
-    // Linux can return EINVAL when write amount is > 0x7ffff000
-    // See https://github.com/ziglang/zig/pull/743#issuecomment-363165856
-    const max_bytes_len = 0x7ffff000;
-
-    var index: usize = 0;
-    while (index < bytes.len) {
-        const amt_to_write = math.min(bytes.len - index, usize(max_bytes_len));
-        const rc = posix.write(fd, bytes.ptr + index, amt_to_write);
-        const write_err = posix.getErrno(rc);
-        switch (write_err) {
-            0 => {
-                index += rc;
-                continue;
-            },
-            posix.EINTR => continue,
-            posix.EINVAL => unreachable,
-            posix.EFAULT => unreachable,
-            posix.EAGAIN => unreachable, // use posixAsyncWrite for non-blocking
-            posix.EBADF => unreachable, // always a race condition
-            posix.EDESTADDRREQ => unreachable, // connect was never called
-            posix.EDQUOT => return PosixWriteError.DiskQuota,
-            posix.EFBIG => return PosixWriteError.FileTooBig,
-            posix.EIO => return PosixWriteError.InputOutput,
-            posix.ENOSPC => return PosixWriteError.NoSpaceLeft,
-            posix.EPERM => return PosixWriteError.AccessDenied,
-            posix.EPIPE => return PosixWriteError.BrokenPipe,
-            else => return unexpectedErrorPosix(write_err),
-        }
-    }
-}
-
-pub fn posix_pwritev(fd: i32, iov: [*]const posix.iovec_const, count: usize, offset: u64) PosixWriteError!void {
-    switch (builtin.os) {
-        builtin.Os.macosx => {
-            // Darwin does not have pwritev but it does have pwrite.
-            var off: usize = 0;
-            var iov_i: usize = 0;
-            var inner_off: usize = 0;
-            while (true) {
-                const v = iov[iov_i];
-                const rc = darwin.pwrite(fd, v.iov_base + inner_off, v.iov_len - inner_off, offset + off);
-                const err = darwin.getErrno(rc);
-                switch (err) {
-                    0 => {
-                        off += rc;
-                        inner_off += rc;
-                        if (inner_off == v.iov_len) {
-                            iov_i += 1;
-                            inner_off = 0;
-                            if (iov_i == count) {
-                                return;
-                            }
-                        }
-                        continue;
-                    },
-                    posix.EINTR => continue,
-                    posix.ESPIPE => unreachable, // fd is not seekable
-                    posix.EINVAL => unreachable,
-                    posix.EFAULT => unreachable,
-                    posix.EAGAIN => unreachable, // use posixAsyncPWriteV for non-blocking
-                    posix.EBADF => unreachable, // always a race condition
-                    posix.EDESTADDRREQ => unreachable, // connect was never called
-                    posix.EDQUOT => return PosixWriteError.DiskQuota,
-                    posix.EFBIG => return PosixWriteError.FileTooBig,
-                    posix.EIO => return PosixWriteError.InputOutput,
-                    posix.ENOSPC => return PosixWriteError.NoSpaceLeft,
-                    posix.EPERM => return PosixWriteError.AccessDenied,
-                    posix.EPIPE => return PosixWriteError.BrokenPipe,
-                    else => return unexpectedErrorPosix(err),
-                }
-            }
-        },
-        builtin.Os.linux, builtin.Os.freebsd, builtin.Os.netbsd => while (true) {
-            const rc = posix.pwritev(fd, iov, count, offset);
-            const err = posix.getErrno(rc);
-            switch (err) {
-                0 => return,
-                posix.EINTR => continue,
-                posix.EINVAL => unreachable,
-                posix.EFAULT => unreachable,
-                posix.EAGAIN => unreachable, // use posixAsyncPWriteV for non-blocking
-                posix.EBADF => unreachable, // always a race condition
-                posix.EDESTADDRREQ => unreachable, // connect was never called
-                posix.EDQUOT => return PosixWriteError.DiskQuota,
-                posix.EFBIG => return PosixWriteError.FileTooBig,
-                posix.EIO => return PosixWriteError.InputOutput,
-                posix.ENOSPC => return PosixWriteError.NoSpaceLeft,
-                posix.EPERM => return PosixWriteError.AccessDenied,
-                posix.EPIPE => return PosixWriteError.BrokenPipe,
-                else => return unexpectedErrorPosix(err),
-            }
-        },
-        else => @compileError("Unsupported OS"),
-    }
-}
-
-pub const PosixOpenError = error{
-    AccessDenied,
-    FileTooBig,
-    IsDir,
-    SymLinkLoop,
-    ProcessFdQuotaExceeded,
-    NameTooLong,
-    SystemFdQuotaExceeded,
-    NoDevice,
-    FileNotFound,
-    SystemResources,
-    NoSpaceLeft,
-    NotDir,
-    PathAlreadyExists,
-    DeviceBusy,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-/// ::file_path needs to be copied in memory to add a null terminating byte.
-/// Calls POSIX open, keeps trying if it gets interrupted, and translates
-/// the return value into zig errors.
-pub fn posixOpen(file_path: []const u8, flags: u32, perm: usize) PosixOpenError!i32 {
-    const file_path_c = try toPosixPath(file_path);
-    return posixOpenC(&file_path_c, flags, perm);
-}
-
-// TODO https://github.com/ziglang/zig/issues/265
-pub fn posixOpenC(file_path: [*]const u8, flags: u32, perm: usize) !i32 {
-    while (true) {
-        const result = posix.open(file_path, flags, perm);
-        const err = posix.getErrno(result);
-        if (err > 0) {
-            switch (err) {
-                posix.EINTR => continue,
-
-                posix.EFAULT => unreachable,
-                posix.EINVAL => unreachable,
-                posix.EACCES => return PosixOpenError.AccessDenied,
-                posix.EFBIG, posix.EOVERFLOW => return PosixOpenError.FileTooBig,
-                posix.EISDIR => return PosixOpenError.IsDir,
-                posix.ELOOP => return PosixOpenError.SymLinkLoop,
-                posix.EMFILE => return PosixOpenError.ProcessFdQuotaExceeded,
-                posix.ENAMETOOLONG => return PosixOpenError.NameTooLong,
-                posix.ENFILE => return PosixOpenError.SystemFdQuotaExceeded,
-                posix.ENODEV => return PosixOpenError.NoDevice,
-                posix.ENOENT => return PosixOpenError.FileNotFound,
-                posix.ENOMEM => return PosixOpenError.SystemResources,
-                posix.ENOSPC => return PosixOpenError.NoSpaceLeft,
-                posix.ENOTDIR => return PosixOpenError.NotDir,
-                posix.EPERM => return PosixOpenError.AccessDenied,
-                posix.EEXIST => return PosixOpenError.PathAlreadyExists,
-                posix.EBUSY => return PosixOpenError.DeviceBusy,
-                else => return unexpectedErrorPosix(err),
-            }
-        }
-        return @intCast(i32, result);
-    }
-}
-
-/// Used to convert a slice to a null terminated slice on the stack.
-/// TODO well defined copy elision
-pub fn toPosixPath(file_path: []const u8) ![posix.PATH_MAX]u8 {
-    var path_with_null: [posix.PATH_MAX]u8 = undefined;
-    if (file_path.len >= posix.PATH_MAX) return error.NameTooLong;
-    mem.copy(u8, path_with_null[0..], file_path);
-    path_with_null[file_path.len] = 0;
-    return path_with_null;
-}
-
-pub fn posixDup2(old_fd: i32, new_fd: i32) !void {
-    while (true) {
-        const err = posix.getErrno(posix.dup2(old_fd, new_fd));
-        if (err > 0) {
-            return switch (err) {
-                posix.EBUSY, posix.EINTR => continue,
-                posix.EMFILE => error.ProcessFdQuotaExceeded,
-                posix.EINVAL => unreachable,
-                else => unexpectedErrorPosix(err),
-            };
-        }
-        return;
-    }
-}
-
-pub fn createNullDelimitedEnvMap(allocator: *Allocator, env_map: *const BufMap) ![]?[*]u8 {
-    const envp_count = env_map.count();
-    const envp_buf = try allocator.alloc(?[*]u8, envp_count + 1);
-    mem.set(?[*]u8, envp_buf, null);
-    errdefer freeNullDelimitedEnvMap(allocator, envp_buf);
-    {
-        var it = env_map.iterator();
-        var i: usize = 0;
-        while (it.next()) |pair| : (i += 1) {
-            const env_buf = try allocator.alloc(u8, pair.key.len + pair.value.len + 2);
-            @memcpy(env_buf.ptr, pair.key.ptr, pair.key.len);
-            env_buf[pair.key.len] = '=';
-            @memcpy(env_buf.ptr + pair.key.len + 1, pair.value.ptr, pair.value.len);
-            env_buf[env_buf.len - 1] = 0;
-
-            envp_buf[i] = env_buf.ptr;
-        }
-        assert(i == envp_count);
-    }
-    assert(envp_buf[envp_count] == null);
-    return envp_buf;
-}
-
-pub fn freeNullDelimitedEnvMap(allocator: *Allocator, envp_buf: []?[*]u8) void {
-    for (envp_buf) |env| {
-        const env_buf = if (env) |ptr| ptr[0 .. cstr.len(ptr) + 1] else break;
-        allocator.free(env_buf);
-    }
-    allocator.free(envp_buf);
-}
-
-/// This function must allocate memory to add a null terminating bytes on path and each arg.
-/// It must also convert to KEY=VALUE\0 format for environment variables, and include null
-/// pointers after the args and after the environment variables.
-/// `argv[0]` is the executable path.
-/// This function also uses the PATH environment variable to get the full path to the executable.
-pub fn posixExecve(argv: []const []const u8, env_map: *const BufMap, allocator: *Allocator) !void {
-    const argv_buf = try allocator.alloc(?[*]u8, argv.len + 1);
-    mem.set(?[*]u8, argv_buf, null);
-    defer {
-        for (argv_buf) |arg| {
-            const arg_buf = if (arg) |ptr| cstr.toSlice(ptr) else break;
-            allocator.free(arg_buf);
-        }
-        allocator.free(argv_buf);
-    }
-    for (argv) |arg, i| {
-        const arg_buf = try allocator.alloc(u8, arg.len + 1);
-        @memcpy(arg_buf.ptr, arg.ptr, arg.len);
-        arg_buf[arg.len] = 0;
-
-        argv_buf[i] = arg_buf.ptr;
-    }
-    argv_buf[argv.len] = null;
-
-    const envp_buf = try createNullDelimitedEnvMap(allocator, env_map);
-    defer freeNullDelimitedEnvMap(allocator, envp_buf);
-
-    const exe_path = argv[0];
-    if (mem.indexOfScalar(u8, exe_path, '/') != null) {
-        return posixExecveErrnoToErr(posix.getErrno(posix.execve(argv_buf[0].?, argv_buf.ptr, envp_buf.ptr)));
-    }
-
-    const PATH = getEnvPosix("PATH") orelse "/usr/local/bin:/bin/:/usr/bin";
-    // PATH.len because it is >= the largest search_path
-    // +1 for the / to join the search path and exe_path
-    // +1 for the null terminating byte
-    const path_buf = try allocator.alloc(u8, PATH.len + exe_path.len + 2);
-    defer allocator.free(path_buf);
-    var it = mem.tokenize(PATH, ":");
-    var seen_eacces = false;
-    var err: usize = undefined;
-    while (it.next()) |search_path| {
-        mem.copy(u8, path_buf, search_path);
-        path_buf[search_path.len] = '/';
-        mem.copy(u8, path_buf[search_path.len + 1 ..], exe_path);
-        path_buf[search_path.len + exe_path.len + 1] = 0;
-        err = posix.getErrno(posix.execve(path_buf.ptr, argv_buf.ptr, envp_buf.ptr));
-        assert(err > 0);
-        if (err == posix.EACCES) {
-            seen_eacces = true;
-        } else if (err != posix.ENOENT) {
-            return posixExecveErrnoToErr(err);
-        }
-    }
-    if (seen_eacces) {
-        err = posix.EACCES;
-    }
-    return posixExecveErrnoToErr(err);
-}
-
-pub const PosixExecveError = error{
-    SystemResources,
-    AccessDenied,
-    InvalidExe,
-    FileSystem,
-    IsDir,
-    FileNotFound,
-    NotDir,
-    FileBusy,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-fn posixExecveErrnoToErr(err: usize) PosixExecveError {
-    assert(err > 0);
-    switch (err) {
-        posix.EFAULT => unreachable,
-        posix.E2BIG => return error.SystemResources,
-        posix.EMFILE => return error.SystemResources,
-        posix.ENAMETOOLONG => return error.SystemResources,
-        posix.ENFILE => return error.SystemResources,
-        posix.ENOMEM => return error.SystemResources,
-        posix.EACCES => return error.AccessDenied,
-        posix.EPERM => return error.AccessDenied,
-        posix.EINVAL => return error.InvalidExe,
-        posix.ENOEXEC => return error.InvalidExe,
-        posix.EIO => return error.FileSystem,
-        posix.ELOOP => return error.FileSystem,
-        posix.EISDIR => return error.IsDir,
-        posix.ENOENT => return error.FileNotFound,
-        posix.ENOTDIR => return error.NotDir,
-        posix.ETXTBSY => return error.FileBusy,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub var linux_elf_aux_maybe: ?[*]std.elf.Auxv = null;
-pub var posix_environ_raw: [][*]u8 = undefined;
-
-/// See std.elf for the constants.
-pub fn linuxGetAuxVal(index: usize) usize {
-    if (builtin.link_libc) {
-        return usize(std.c.getauxval(index));
-    } else if (linux_elf_aux_maybe) |auxv| {
-        var i: usize = 0;
-        while (auxv[i].a_type != std.elf.AT_NULL) : (i += 1) {
-            if (auxv[i].a_type == index)
-                return auxv[i].a_un.a_val;
-        }
-    }
-    return 0;
-}
 
 pub fn getBaseAddress() usize {
     switch (builtin.os) {
@@ -803,7 +215,7 @@ pub fn getEnvMap(allocator: *Allocator) !BufMap {
         }
         return result;
     } else {
-        for (posix_environ_raw) |ptr| {
+        for (posix.environ) |ptr| {
             var line_i: usize = 0;
             while (ptr[line_i] != 0 and ptr[line_i] != '=') : (line_i += 1) {}
             const key = ptr[0..line_i];
@@ -821,23 +233,6 @@ pub fn getEnvMap(allocator: *Allocator) !BufMap {
 test "os.getEnvMap" {
     var env = try getEnvMap(std.debug.global_allocator);
     defer env.deinit();
-}
-
-/// TODO make this go through libc when we have it
-pub fn getEnvPosix(key: []const u8) ?[]const u8 {
-    for (posix_environ_raw) |ptr| {
-        var line_i: usize = 0;
-        while (ptr[line_i] != 0 and ptr[line_i] != '=') : (line_i += 1) {}
-        const this_key = ptr[0..line_i];
-        if (!mem.eql(u8, key, this_key)) continue;
-
-        var end_i: usize = line_i;
-        while (ptr[end_i] != 0) : (end_i += 1) {}
-        const this_value = ptr[line_i + 1 .. end_i];
-
-        return this_value;
-    }
-    return null;
 }
 
 pub const GetEnvVarOwnedError = error{
@@ -896,130 +291,22 @@ test "os.getEnvVarOwned" {
     testing.expectError(error.EnvironmentVariableNotFound, getEnvVarOwned(ga, "BADENV"));
 }
 
+/// The result is a slice of `out_buffer`, from index `0`.
+pub fn getCwd(out_buffer: *[MAX_PATH_BYTES]u8) ![]u8 {
+    return posix.getcwd(out_buffer);
+}
+
 /// Caller must free the returned memory.
 pub fn getCwdAlloc(allocator: *Allocator) ![]u8 {
-    var buf: [MAX_PATH_BYTES]u8 = undefined;
-    return mem.dupe(allocator, u8, try getCwd(&buf));
+    var buf: [os.MAX_PATH_BYTES]u8 = undefined;
+    return mem.dupe(allocator, u8, try posix.getcwd(&buf));
 }
 
-pub const GetCwdError = error{Unexpected};
-
-/// The result is a slice of out_buffer.
-pub fn getCwd(out_buffer: *[MAX_PATH_BYTES]u8) GetCwdError![]u8 {
-    switch (builtin.os) {
-        Os.windows => {
-            var utf16le_buf: [windows_util.PATH_MAX_WIDE]u16 = undefined;
-            const casted_len = @intCast(windows.DWORD, utf16le_buf.len); // TODO shouldn't need this cast
-            const casted_ptr = ([*]u16)(&utf16le_buf); // TODO shouldn't need this cast
-            const result = windows.GetCurrentDirectoryW(casted_len, casted_ptr);
-            if (result == 0) {
-                const err = windows.GetLastError();
-                switch (err) {
-                    else => return unexpectedErrorWindows(err),
-                }
-            }
-            assert(result <= utf16le_buf.len);
-            const utf16le_slice = utf16le_buf[0..result];
-            // Trust that Windows gives us valid UTF-16LE.
-            const end_index = std.unicode.utf16leToUtf8(out_buffer, utf16le_slice) catch unreachable;
-            return out_buffer[0..end_index];
-        },
-        else => {
-            const err = posix.getErrno(posix.getcwd(out_buffer, out_buffer.len));
-            switch (err) {
-                0 => return cstr.toSlice(out_buffer),
-                posix.ERANGE => unreachable,
-                else => return unexpectedErrorPosix(err),
-            }
-        },
-    }
-}
-
-test "os.getCwd" {
+test "getCwdAlloc" {
     // at least call it so it gets compiled
-    _ = getCwdAlloc(debug.global_allocator) catch undefined;
-    var buf: [MAX_PATH_BYTES]u8 = undefined;
-    _ = getCwd(&buf) catch undefined;
-}
-
-pub const SymLinkError = PosixSymLinkError || WindowsSymLinkError;
-
-/// TODO add a symLinkC variant
-pub fn symLink(existing_path: []const u8, new_path: []const u8) SymLinkError!void {
-    if (is_windows) {
-        return symLinkWindows(existing_path, new_path);
-    } else {
-        return symLinkPosix(existing_path, new_path);
-    }
-}
-
-pub const WindowsSymLinkError = error{
-    NameTooLong,
-    InvalidUtf8,
-    BadPathName,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn symLinkW(existing_path_w: [*]const u16, new_path_w: [*]const u16) WindowsSymLinkError!void {
-    if (windows.CreateSymbolicLinkW(existing_path_w, new_path_w, 0) == 0) {
-        const err = windows.GetLastError();
-        switch (err) {
-            else => return unexpectedErrorWindows(err),
-        }
-    }
-}
-
-pub fn symLinkWindows(existing_path: []const u8, new_path: []const u8) WindowsSymLinkError!void {
-    const existing_path_w = try windows_util.sliceToPrefixedFileW(existing_path);
-    const new_path_w = try windows_util.sliceToPrefixedFileW(new_path);
-    return symLinkW(&existing_path_w, &new_path_w);
-}
-
-pub const PosixSymLinkError = error{
-    AccessDenied,
-    DiskQuota,
-    PathAlreadyExists,
-    FileSystem,
-    SymLinkLoop,
-    NameTooLong,
-    FileNotFound,
-    SystemResources,
-    NoSpaceLeft,
-    ReadOnlyFileSystem,
-    NotDir,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn symLinkPosixC(existing_path: [*]const u8, new_path: [*]const u8) PosixSymLinkError!void {
-    const err = posix.getErrno(posix.symlink(existing_path, new_path));
-    switch (err) {
-        0 => return,
-        posix.EFAULT => unreachable,
-        posix.EINVAL => unreachable,
-        posix.EACCES => return error.AccessDenied,
-        posix.EPERM => return error.AccessDenied,
-        posix.EDQUOT => return error.DiskQuota,
-        posix.EEXIST => return error.PathAlreadyExists,
-        posix.EIO => return error.FileSystem,
-        posix.ELOOP => return error.SymLinkLoop,
-        posix.ENAMETOOLONG => return error.NameTooLong,
-        posix.ENOENT => return error.FileNotFound,
-        posix.ENOTDIR => return error.NotDir,
-        posix.ENOMEM => return error.SystemResources,
-        posix.ENOSPC => return error.NoSpaceLeft,
-        posix.EROFS => return error.ReadOnlyFileSystem,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub fn symLinkPosix(existing_path: []const u8, new_path: []const u8) PosixSymLinkError!void {
-    const existing_path_c = try toPosixPath(existing_path);
-    const new_path_c = try toPosixPath(new_path);
-    return symLinkPosixC(&existing_path_c, &new_path_c);
+    var buf: [1000]u8 = undefined;
+    const allocator = &std.heap.FixedBufferAllocator.init(&buf).allocator;
+    _ = getCwdAlloc(allocator) catch {};
 }
 
 // here we replace the standard +/ with -_ so that it can be used in a file name
@@ -1050,78 +337,6 @@ pub fn atomicSymLink(allocator: *Allocator, existing_path: []const u8, new_path:
         } else |err| switch (err) {
             error.PathAlreadyExists => continue,
             else => return err, // TODO zig should know this set does not include PathAlreadyExists
-        }
-    }
-}
-
-pub const DeleteFileError = error{
-    FileNotFound,
-    AccessDenied,
-    FileBusy,
-    FileSystem,
-    IsDir,
-    SymLinkLoop,
-    NameTooLong,
-    NotDir,
-    SystemResources,
-    ReadOnlyFileSystem,
-
-    /// On Windows, file paths must be valid Unicode.
-    InvalidUtf8,
-
-    /// On Windows, file paths cannot contain these characters:
-    /// '/', '*', '?', '"', '<', '>', '|'
-    BadPathName,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn deleteFile(file_path: []const u8) DeleteFileError!void {
-    if (builtin.os == Os.windows) {
-        const file_path_w = try windows_util.sliceToPrefixedFileW(file_path);
-        return deleteFileW(&file_path_w);
-    } else {
-        const file_path_c = try toPosixPath(file_path);
-        return deleteFileC(&file_path_c);
-    }
-}
-
-pub fn deleteFileW(file_path: [*]const u16) DeleteFileError!void {
-    if (windows.DeleteFileW(file_path) == 0) {
-        const err = windows.GetLastError();
-        switch (err) {
-            windows.ERROR.FILE_NOT_FOUND => return error.FileNotFound,
-            windows.ERROR.ACCESS_DENIED => return error.AccessDenied,
-            windows.ERROR.FILENAME_EXCED_RANGE => return error.NameTooLong,
-            windows.ERROR.INVALID_PARAMETER => return error.NameTooLong,
-            else => return unexpectedErrorWindows(err),
-        }
-    }
-}
-
-pub fn deleteFileC(file_path: [*]const u8) DeleteFileError!void {
-    if (is_windows) {
-        const file_path_w = try windows_util.cStrToPrefixedFileW(file_path);
-        return deleteFileW(&file_path_w);
-    } else {
-        const err = posix.getErrno(posix.unlink(file_path));
-        switch (err) {
-            0 => return,
-            posix.EACCES => return error.AccessDenied,
-            posix.EPERM => return error.AccessDenied,
-            posix.EBUSY => return error.FileBusy,
-            posix.EFAULT => unreachable,
-            posix.EINVAL => unreachable,
-            posix.EIO => return error.FileSystem,
-            posix.EISDIR => return error.IsDir,
-            posix.ELOOP => return error.SymLinkLoop,
-            posix.ENAMETOOLONG => return error.NameTooLong,
-            posix.ENOENT => return error.FileNotFound,
-            posix.ENOTDIR => return error.NotDir,
-            posix.ENOMEM => return error.SystemResources,
-            posix.EROFS => return error.ReadOnlyFileSystem,
-            else => return unexpectedErrorPosix(err),
         }
     }
 }
@@ -1236,8 +451,8 @@ pub const AtomicFile = struct {
             const dest_path_c = try toPosixPath(self.dest_path);
             return renameC(&self.tmp_path_buf, &dest_path_c);
         } else if (is_windows) {
-            const dest_path_w = try windows_util.sliceToPrefixedFileW(self.dest_path);
-            const tmp_path_w = try windows_util.cStrToPrefixedFileW(&self.tmp_path_buf);
+            const dest_path_w = try posix.sliceToPrefixedFileW(self.dest_path);
+            const tmp_path_w = try posix.cStrToPrefixedFileW(&self.tmp_path_buf);
             return renameW(&tmp_path_w, &dest_path_w);
         } else {
             @compileError("Unsupported OS");
@@ -1245,109 +460,27 @@ pub const AtomicFile = struct {
     }
 };
 
-pub fn renameC(old_path: [*]const u8, new_path: [*]const u8) !void {
-    if (is_windows) {
-        const old_path_w = try windows_util.cStrToPrefixedFileW(old_path);
-        const new_path_w = try windows_util.cStrToPrefixedFileW(new_path);
-        return renameW(&old_path_w, &new_path_w);
-    } else {
-        const err = posix.getErrno(posix.rename(old_path, new_path));
-        switch (err) {
-            0 => return,
-            posix.EACCES => return error.AccessDenied,
-            posix.EPERM => return error.AccessDenied,
-            posix.EBUSY => return error.FileBusy,
-            posix.EDQUOT => return error.DiskQuota,
-            posix.EFAULT => unreachable,
-            posix.EINVAL => unreachable,
-            posix.EISDIR => return error.IsDir,
-            posix.ELOOP => return error.SymLinkLoop,
-            posix.EMLINK => return error.LinkQuotaExceeded,
-            posix.ENAMETOOLONG => return error.NameTooLong,
-            posix.ENOENT => return error.FileNotFound,
-            posix.ENOTDIR => return error.NotDir,
-            posix.ENOMEM => return error.SystemResources,
-            posix.ENOSPC => return error.NoSpaceLeft,
-            posix.EEXIST => return error.PathAlreadyExists,
-            posix.ENOTEMPTY => return error.PathAlreadyExists,
-            posix.EROFS => return error.ReadOnlyFileSystem,
-            posix.EXDEV => return error.RenameAcrossMountPoints,
-            else => return unexpectedErrorPosix(err),
-        }
-    }
-}
+const default_new_dir_mode = 0o755;
 
-pub fn renameW(old_path: [*]const u16, new_path: [*]const u16) !void {
-    const flags = windows.MOVEFILE_REPLACE_EXISTING | windows.MOVEFILE_WRITE_THROUGH;
-    if (windows.MoveFileExW(old_path, new_path, flags) == 0) {
-        const err = windows.GetLastError();
-        switch (err) {
-            else => return unexpectedErrorWindows(err),
-        }
-    }
-}
-
-pub fn rename(old_path: []const u8, new_path: []const u8) !void {
-    if (is_windows) {
-        const old_path_w = try windows_util.sliceToPrefixedFileW(old_path);
-        const new_path_w = try windows_util.sliceToPrefixedFileW(new_path);
-        return renameW(&old_path_w, &new_path_w);
-    } else {
-        const old_path_c = try toPosixPath(old_path);
-        const new_path_c = try toPosixPath(new_path);
-        return renameC(&old_path_c, &new_path_c);
-    }
-}
-
+/// Create a new directory.
 pub fn makeDir(dir_path: []const u8) !void {
-    if (is_windows) {
-        return makeDirWindows(dir_path);
-    } else {
-        return makeDirPosix(dir_path);
-    }
+    return posix.mkdir(dir_path, default_new_dir_mode);
 }
 
-pub fn makeDirWindows(dir_path: []const u8) !void {
-    const dir_path_w = try windows_util.sliceToPrefixedFileW(dir_path);
-
-    if (windows.CreateDirectoryW(&dir_path_w, null) == 0) {
-        const err = windows.GetLastError();
-        return switch (err) {
-            windows.ERROR.ALREADY_EXISTS => error.PathAlreadyExists,
-            windows.ERROR.PATH_NOT_FOUND => error.FileNotFound,
-            else => unexpectedErrorWindows(err),
-        };
-    }
+/// Same as `makeDir` except the parameter is a null-terminated UTF8-encoded string.
+pub fn makeDirC(dir_path: [*]const u8) !void {
+    return posix.mkdirC(dir_path, default_new_dir_mode);
 }
 
-pub fn makeDirPosixC(dir_path: [*]const u8) !void {
-    const err = posix.getErrno(posix.mkdir(dir_path, 0o755));
-    switch (err) {
-        0 => return,
-        posix.EACCES => return error.AccessDenied,
-        posix.EPERM => return error.AccessDenied,
-        posix.EDQUOT => return error.DiskQuota,
-        posix.EEXIST => return error.PathAlreadyExists,
-        posix.EFAULT => unreachable,
-        posix.ELOOP => return error.SymLinkLoop,
-        posix.EMLINK => return error.LinkQuotaExceeded,
-        posix.ENAMETOOLONG => return error.NameTooLong,
-        posix.ENOENT => return error.FileNotFound,
-        posix.ENOMEM => return error.SystemResources,
-        posix.ENOSPC => return error.NoSpaceLeft,
-        posix.ENOTDIR => return error.NotDir,
-        posix.EROFS => return error.ReadOnlyFileSystem,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub fn makeDirPosix(dir_path: []const u8) !void {
-    const dir_path_c = try toPosixPath(dir_path);
-    return makeDirPosixC(&dir_path_c);
+/// Same as `makeDir` except the parameter is a null-terminated UTF16LE-encoded string.
+pub fn makeDirW(dir_path: [*]const u16) !void {
+    return posix.mkdirW(dir_path, default_new_dir_mode);
 }
 
 /// Calls makeDir recursively to make an entire path. Returns success if the path
 /// already exists and is a directory.
+/// This function is not atomic, and if it returns an error, the file system may
+/// have been modified regardless.
 /// TODO determine if we can remove the allocator requirement from this function
 pub fn makePath(allocator: *Allocator, full_path: []const u8) !void {
     const resolved_path = try path.resolve(allocator, [][]const u8{full_path});
@@ -1381,78 +514,20 @@ pub fn makePath(allocator: *Allocator, full_path: []const u8) !void {
     }
 }
 
-pub const DeleteDirError = error{
-    AccessDenied,
-    FileBusy,
-    SymLinkLoop,
-    NameTooLong,
-    FileNotFound,
-    SystemResources,
-    NotDir,
-    DirNotEmpty,
-    ReadOnlyFileSystem,
-    InvalidUtf8,
-    BadPathName,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn deleteDirC(dir_path: [*]const u8) DeleteDirError!void {
-    switch (builtin.os) {
-        Os.windows => {
-            const dir_path_w = try windows_util.cStrToPrefixedFileW(dir_path);
-            return deleteDirW(&dir_path_w);
-        },
-        Os.linux, Os.macosx, Os.ios, Os.freebsd, Os.netbsd => {
-            const err = posix.getErrno(posix.rmdir(dir_path));
-            switch (err) {
-                0 => return,
-                posix.EACCES => return error.AccessDenied,
-                posix.EPERM => return error.AccessDenied,
-                posix.EBUSY => return error.FileBusy,
-                posix.EFAULT => unreachable,
-                posix.EINVAL => unreachable,
-                posix.ELOOP => return error.SymLinkLoop,
-                posix.ENAMETOOLONG => return error.NameTooLong,
-                posix.ENOENT => return error.FileNotFound,
-                posix.ENOMEM => return error.SystemResources,
-                posix.ENOTDIR => return error.NotDir,
-                posix.EEXIST => return error.DirNotEmpty,
-                posix.ENOTEMPTY => return error.DirNotEmpty,
-                posix.EROFS => return error.ReadOnlyFileSystem,
-                else => return unexpectedErrorPosix(err),
-            }
-        },
-        else => @compileError("unimplemented"),
-    }
-}
-
-pub fn deleteDirW(dir_path_w: [*]const u16) DeleteDirError!void {
-    if (windows.RemoveDirectoryW(dir_path_w) == 0) {
-        const err = windows.GetLastError();
-        switch (err) {
-            windows.ERROR.PATH_NOT_FOUND => return error.FileNotFound,
-            windows.ERROR.DIR_NOT_EMPTY => return error.DirNotEmpty,
-            else => return unexpectedErrorWindows(err),
-        }
-    }
-}
-
-/// Returns ::error.DirNotEmpty if the directory is not empty.
-/// To delete a directory recursively, see ::deleteTree
+/// Returns `error.DirNotEmpty` if the directory is not empty.
+/// To delete a directory recursively, see `deleteTree`.
 pub fn deleteDir(dir_path: []const u8) DeleteDirError!void {
-    switch (builtin.os) {
-        Os.windows => {
-            const dir_path_w = try windows_util.sliceToPrefixedFileW(dir_path);
-            return deleteDirW(&dir_path_w);
-        },
-        Os.linux, Os.macosx, Os.ios, Os.freebsd, Os.netbsd => {
-            const dir_path_c = try toPosixPath(dir_path);
-            return deleteDirC(&dir_path_c);
-        },
-        else => @compileError("unimplemented"),
-    }
+    return posix.rmdir(dir_path);
+}
+
+/// Same as `deleteDir` except the parameter is a null-terminated UTF8-encoded string.
+pub fn deleteDirC(dir_path: [*]const u8) DeleteDirError!void {
+    return posix.rmdirC(dir_path);
+}
+
+/// Same as `deleteDir` except the parameter is a null-terminated UTF16LE-encoded string.
+pub fn deleteDirW(dir_path: [*]const u16) DeleteDirError!void {
+    return posix.rmdirW(dir_path);
 }
 
 /// Whether ::full_path describes a symlink, file, or directory, this function
@@ -1486,7 +561,6 @@ const DeleteTreeError = error{
     /// '/', '*', '?', '"', '<', '>', '|'
     BadPathName,
 
-    /// See https://github.com/ziglang/zig/issues/1396
     Unexpected,
 };
 
@@ -1624,7 +698,6 @@ pub const Dir = struct {
         BadPathName,
         DeviceBusy,
 
-        /// See https://github.com/ziglang/zig/issues/1396
         Unexpected,
     };
 
@@ -1886,113 +959,15 @@ pub const Dir = struct {
     }
 };
 
-pub fn changeCurDir(dir_path: []const u8) !void {
-    const dir_path_c = try toPosixPath(dir_path);
-    const err = posix.getErrno(posix.chdir(&dir_path_c));
-    switch (err) {
-        0 => return,
-        posix.EACCES => return error.AccessDenied,
-        posix.EFAULT => unreachable,
-        posix.EIO => return error.FileSystem,
-        posix.ELOOP => return error.SymLinkLoop,
-        posix.ENAMETOOLONG => return error.NameTooLong,
-        posix.ENOENT => return error.FileNotFound,
-        posix.ENOMEM => return error.SystemResources,
-        posix.ENOTDIR => return error.NotDir,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
 /// Read value of a symbolic link.
-/// The return value is a slice of out_buffer.
-pub fn readLinkC(out_buffer: *[posix.PATH_MAX]u8, pathname: [*]const u8) ![]u8 {
-    const rc = posix.readlink(pathname, out_buffer, out_buffer.len);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return out_buffer[0..rc],
-        posix.EACCES => return error.AccessDenied,
-        posix.EFAULT => unreachable,
-        posix.EINVAL => unreachable,
-        posix.EIO => return error.FileSystem,
-        posix.ELOOP => return error.SymLinkLoop,
-        posix.ENAMETOOLONG => unreachable, // out_buffer is at least PATH_MAX
-        posix.ENOENT => return error.FileNotFound,
-        posix.ENOMEM => return error.SystemResources,
-        posix.ENOTDIR => return error.NotDir,
-        else => return unexpectedErrorPosix(err),
-    }
+/// The return value is a slice of buffer, from index `0`.
+pub fn readLink(buffer: *[posix.PATH_MAX]u8, pathname: []const u8) ![]u8 {
+    return posix.readlink(pathname, buffer);
 }
 
-/// Read value of a symbolic link.
-/// The return value is a slice of out_buffer.
-pub fn readLink(out_buffer: *[posix.PATH_MAX]u8, file_path: []const u8) ![]u8 {
-    const file_path_c = try toPosixPath(file_path);
-    return readLinkC(out_buffer, &file_path_c);
-}
-
-pub fn posix_setuid(uid: u32) !void {
-    const err = posix.getErrno(posix.setuid(uid));
-    if (err == 0) return;
-    return switch (err) {
-        posix.EAGAIN => error.ResourceLimitReached,
-        posix.EINVAL => error.InvalidUserId,
-        posix.EPERM => error.PermissionDenied,
-        else => unexpectedErrorPosix(err),
-    };
-}
-
-pub fn posix_setreuid(ruid: u32, euid: u32) !void {
-    const err = posix.getErrno(posix.setreuid(ruid, euid));
-    if (err == 0) return;
-    return switch (err) {
-        posix.EAGAIN => error.ResourceLimitReached,
-        posix.EINVAL => error.InvalidUserId,
-        posix.EPERM => error.PermissionDenied,
-        else => unexpectedErrorPosix(err),
-    };
-}
-
-pub fn posix_setgid(gid: u32) !void {
-    const err = posix.getErrno(posix.setgid(gid));
-    if (err == 0) return;
-    return switch (err) {
-        posix.EAGAIN => error.ResourceLimitReached,
-        posix.EINVAL => error.InvalidUserId,
-        posix.EPERM => error.PermissionDenied,
-        else => unexpectedErrorPosix(err),
-    };
-}
-
-pub fn posix_setregid(rgid: u32, egid: u32) !void {
-    const err = posix.getErrno(posix.setregid(rgid, egid));
-    if (err == 0) return;
-    return switch (err) {
-        posix.EAGAIN => error.ResourceLimitReached,
-        posix.EINVAL => error.InvalidUserId,
-        posix.EPERM => error.PermissionDenied,
-        else => unexpectedErrorPosix(err),
-    };
-}
-
-pub const WindowsGetStdHandleErrs = error{
-    NoStdHandles,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn windowsGetStdHandle(handle_id: windows.DWORD) WindowsGetStdHandleErrs!windows.HANDLE {
-    if (windows.GetStdHandle(handle_id)) |handle| {
-        if (handle == windows.INVALID_HANDLE_VALUE) {
-            const err = windows.GetLastError();
-            return switch (err) {
-                else => os.unexpectedErrorWindows(err),
-            };
-        }
-        return handle;
-    } else {
-        return error.NoStdHandles;
-    }
+/// Same as `readLink`, except the `pathname` parameter is null-terminated.
+pub fn readLinkC(buffer: *[posix.PATH_MAX]u8, pathname: [*]const u8) ![]u8 {
+    return posix.readlinkC(pathname, buffer);
 }
 
 pub const ArgIteratorPosix = struct {
@@ -2328,34 +1303,6 @@ fn testWindowsCmdLine(input_cmd_line: [*]const u8, expected_args: []const []cons
     testing.expect(it.next(debug.global_allocator) == null);
 }
 
-// TODO make this a build variable that you can set
-const unexpected_error_tracing = false;
-const UnexpectedError = error{
-    /// The Operating System returned an undocumented error code.
-    Unexpected,
-};
-
-/// Call this when you made a syscall or something that sets errno
-/// and you get an unexpected error.
-pub fn unexpectedErrorPosix(errno: usize) UnexpectedError {
-    if (unexpected_error_tracing) {
-        debug.warn("unexpected errno: {}\n", errno);
-        debug.dumpCurrentStackTrace(null);
-    }
-    return error.Unexpected;
-}
-
-/// Call this when you made a windows DLL call or something that does SetLastError
-/// and you get an unexpected error.
-pub fn unexpectedErrorWindows(err: windows.DWORD) UnexpectedError {
-    if (unexpected_error_tracing) {
-        debug.warn("unexpected GetLastError(): {}\n", err);
-        @breakpoint();
-        debug.dumpCurrentStackTrace(null);
-    }
-    return error.Unexpected;
-}
-
 pub fn openSelfExe() !os.File {
     switch (builtin.os) {
         Os.linux => return os.File.openReadC(c"/proc/self/exe"),
@@ -2366,7 +1313,7 @@ pub fn openSelfExe() !os.File {
             return os.File.openReadC(self_exe_path.ptr);
         },
         Os.windows => {
-            var buf: [windows_util.PATH_MAX_WIDE]u16 = undefined;
+            var buf: [posix.PATH_MAX_WIDE]u16 = undefined;
             const wide_slice = try selfExePathW(&buf);
             return os.File.openReadW(wide_slice.ptr);
         },
@@ -2381,7 +1328,7 @@ test "openSelfExe" {
     }
 }
 
-pub fn selfExePathW(out_buffer: *[windows_util.PATH_MAX_WIDE]u16) ![]u16 {
+pub fn selfExePathW(out_buffer: *[posix.PATH_MAX_WIDE]u16) ![]u16 {
     const casted_len = @intCast(windows.DWORD, out_buffer.len); // TODO shouldn't need this cast
     const rc = windows.GetModuleFileNameW(null, out_buffer, casted_len);
     assert(rc <= out_buffer.len);
@@ -2434,7 +1381,7 @@ pub fn selfExePath(out_buffer: *[MAX_PATH_BYTES]u8) ![]u8 {
             };
         },
         Os.windows => {
-            var utf16le_buf: [windows_util.PATH_MAX_WIDE]u16 = undefined;
+            var utf16le_buf: [posix.PATH_MAX_WIDE]u16 = undefined;
             const utf16le_slice = try selfExePathW(&utf16le_buf);
             // Trust that Windows gives us valid UTF-16LE.
             const end_index = std.unicode.utf16leToUtf8(out_buffer, utf16le_slice) catch unreachable;
@@ -2478,521 +1425,6 @@ pub fn selfExeDirPath(out_buffer: *[MAX_PATH_BYTES]u8) ![]const u8 {
             return path.dirname(self_exe_path).?;
         },
         else => @compileError("Unsupported OS"),
-    }
-}
-
-pub fn isTty(handle: FileHandle) bool {
-    if (is_windows) {
-        return windows_util.windowsIsTty(handle);
-    } else {
-        if (builtin.link_libc) {
-            return c.isatty(handle) != 0;
-        } else {
-            return posix.isatty(handle);
-        }
-    }
-}
-
-pub fn supportsAnsiEscapeCodes(handle: FileHandle) bool {
-    if (is_windows) {
-        return windows_util.windowsIsCygwinPty(handle);
-    } else {
-        if (builtin.link_libc) {
-            return c.isatty(handle) != 0;
-        } else {
-            return posix.isatty(handle);
-        }
-    }
-}
-
-pub const PosixSocketError = error{
-    /// Permission to create a socket of the specified type and/or
-    /// pro‐tocol is denied.
-    PermissionDenied,
-
-    /// The implementation does not support the specified address family.
-    AddressFamilyNotSupported,
-
-    /// Unknown protocol, or protocol family not available.
-    ProtocolFamilyNotAvailable,
-
-    /// The per-process limit on the number of open file descriptors has been reached.
-    ProcessFdQuotaExceeded,
-
-    /// The system-wide limit on the total number of open files has been reached.
-    SystemFdQuotaExceeded,
-
-    /// Insufficient memory is available. The socket cannot be created until sufficient
-    /// resources are freed.
-    SystemResources,
-
-    /// The protocol type or the specified protocol is not supported within this domain.
-    ProtocolNotSupported,
-};
-
-pub fn posixSocket(domain: u32, socket_type: u32, protocol: u32) !i32 {
-    const rc = posix.socket(domain, socket_type, protocol);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return @intCast(i32, rc),
-        posix.EACCES => return PosixSocketError.PermissionDenied,
-        posix.EAFNOSUPPORT => return PosixSocketError.AddressFamilyNotSupported,
-        posix.EINVAL => return PosixSocketError.ProtocolFamilyNotAvailable,
-        posix.EMFILE => return PosixSocketError.ProcessFdQuotaExceeded,
-        posix.ENFILE => return PosixSocketError.SystemFdQuotaExceeded,
-        posix.ENOBUFS, posix.ENOMEM => return PosixSocketError.SystemResources,
-        posix.EPROTONOSUPPORT => return PosixSocketError.ProtocolNotSupported,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub const PosixBindError = error{
-    /// The address is protected, and the user is not the superuser.
-    /// For UNIX domain sockets: Search permission is denied on  a  component
-    /// of  the  path  prefix.
-    AccessDenied,
-
-    /// The given address is already in use, or in the case of Internet domain sockets,
-    /// The  port number was specified as zero in the socket
-    /// address structure, but, upon attempting to bind to  an  ephemeral  port,  it  was
-    /// determined  that  all  port  numbers in the ephemeral port range are currently in
-    /// use.  See the discussion of /proc/sys/net/ipv4/ip_local_port_range ip(7).
-    AddressInUse,
-
-    /// A nonexistent interface was requested or the requested address was not local.
-    AddressNotAvailable,
-
-    /// Too many symbolic links were encountered in resolving addr.
-    SymLinkLoop,
-
-    /// addr is too long.
-    NameTooLong,
-
-    /// A component in the directory prefix of the socket pathname does not exist.
-    FileNotFound,
-
-    /// Insufficient kernel memory was available.
-    SystemResources,
-
-    /// A component of the path prefix is not a directory.
-    NotDir,
-
-    /// The socket inode would reside on a read-only filesystem.
-    ReadOnlyFileSystem,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-/// addr is `&const T` where T is one of the sockaddr
-pub fn posixBind(fd: i32, addr: *const posix.sockaddr) PosixBindError!void {
-    const rc = posix.bind(fd, addr, @sizeOf(posix.sockaddr));
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return,
-        posix.EACCES => return PosixBindError.AccessDenied,
-        posix.EADDRINUSE => return PosixBindError.AddressInUse,
-        posix.EBADF => unreachable, // always a race condition if this error is returned
-        posix.EINVAL => unreachable,
-        posix.ENOTSOCK => unreachable,
-        posix.EADDRNOTAVAIL => return PosixBindError.AddressNotAvailable,
-        posix.EFAULT => unreachable,
-        posix.ELOOP => return PosixBindError.SymLinkLoop,
-        posix.ENAMETOOLONG => return PosixBindError.NameTooLong,
-        posix.ENOENT => return PosixBindError.FileNotFound,
-        posix.ENOMEM => return PosixBindError.SystemResources,
-        posix.ENOTDIR => return PosixBindError.NotDir,
-        posix.EROFS => return PosixBindError.ReadOnlyFileSystem,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-const PosixListenError = error{
-    /// Another socket is already listening on the same port.
-    /// For Internet domain sockets, the  socket referred to by sockfd had not previously
-    /// been bound to an address and, upon attempting to bind it to an ephemeral port, it
-    /// was determined that all port numbers in the ephemeral port range are currently in
-    /// use.  See the discussion of /proc/sys/net/ipv4/ip_local_port_range in ip(7).
-    AddressInUse,
-
-    /// The file descriptor sockfd does not refer to a socket.
-    FileDescriptorNotASocket,
-
-    /// The socket is not of a type that supports the listen() operation.
-    OperationNotSupported,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn posixListen(sockfd: i32, backlog: u32) PosixListenError!void {
-    const rc = posix.listen(sockfd, backlog);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return,
-        posix.EADDRINUSE => return PosixListenError.AddressInUse,
-        posix.EBADF => unreachable,
-        posix.ENOTSOCK => return PosixListenError.FileDescriptorNotASocket,
-        posix.EOPNOTSUPP => return PosixListenError.OperationNotSupported,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub const PosixAcceptError = error{
-    ConnectionAborted,
-
-    /// The per-process limit on the number of open file descriptors has been reached.
-    ProcessFdQuotaExceeded,
-
-    /// The system-wide limit on the total number of open files has been reached.
-    SystemFdQuotaExceeded,
-
-    /// Not enough free memory.  This often means that the memory allocation  is  limited
-    /// by the socket buffer limits, not by the system memory.
-    SystemResources,
-
-    /// The file descriptor sockfd does not refer to a socket.
-    FileDescriptorNotASocket,
-
-    /// The referenced socket is not of type SOCK_STREAM.
-    OperationNotSupported,
-
-    ProtocolFailure,
-
-    /// Firewall rules forbid connection.
-    BlockedByFirewall,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn posixAccept(fd: i32, addr: *posix.sockaddr, flags: u32) PosixAcceptError!i32 {
-    while (true) {
-        var sockaddr_size = u32(@sizeOf(posix.sockaddr));
-        const rc = posix.accept4(fd, addr, &sockaddr_size, flags);
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0 => return @intCast(i32, rc),
-            posix.EINTR => continue,
-            else => return unexpectedErrorPosix(err),
-
-            posix.EAGAIN => unreachable, // use posixAsyncAccept for non-blocking
-            posix.EBADF => unreachable, // always a race condition
-            posix.ECONNABORTED => return PosixAcceptError.ConnectionAborted,
-            posix.EFAULT => unreachable,
-            posix.EINVAL => unreachable,
-            posix.EMFILE => return PosixAcceptError.ProcessFdQuotaExceeded,
-            posix.ENFILE => return PosixAcceptError.SystemFdQuotaExceeded,
-            posix.ENOBUFS => return PosixAcceptError.SystemResources,
-            posix.ENOMEM => return PosixAcceptError.SystemResources,
-            posix.ENOTSOCK => return PosixAcceptError.FileDescriptorNotASocket,
-            posix.EOPNOTSUPP => return PosixAcceptError.OperationNotSupported,
-            posix.EPROTO => return PosixAcceptError.ProtocolFailure,
-            posix.EPERM => return PosixAcceptError.BlockedByFirewall,
-        }
-    }
-}
-
-/// Returns -1 if would block.
-pub fn posixAsyncAccept(fd: i32, addr: *posix.sockaddr, flags: u32) PosixAcceptError!i32 {
-    while (true) {
-        var sockaddr_size = u32(@sizeOf(posix.sockaddr));
-        const rc = posix.accept4(fd, addr, &sockaddr_size, flags);
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0 => return @intCast(i32, rc),
-            posix.EINTR => continue,
-            else => return unexpectedErrorPosix(err),
-
-            posix.EAGAIN => return -1,
-            posix.EBADF => unreachable, // always a race condition
-            posix.ECONNABORTED => return PosixAcceptError.ConnectionAborted,
-            posix.EFAULT => unreachable,
-            posix.EINVAL => unreachable,
-            posix.EMFILE => return PosixAcceptError.ProcessFdQuotaExceeded,
-            posix.ENFILE => return PosixAcceptError.SystemFdQuotaExceeded,
-            posix.ENOBUFS => return PosixAcceptError.SystemResources,
-            posix.ENOMEM => return PosixAcceptError.SystemResources,
-            posix.ENOTSOCK => return PosixAcceptError.FileDescriptorNotASocket,
-            posix.EOPNOTSUPP => return PosixAcceptError.OperationNotSupported,
-            posix.EPROTO => return PosixAcceptError.ProtocolFailure,
-            posix.EPERM => return PosixAcceptError.BlockedByFirewall,
-        }
-    }
-}
-
-pub const LinuxEpollCreateError = error{
-    /// The  per-user   limit   on   the   number   of   epoll   instances   imposed   by
-    /// /proc/sys/fs/epoll/max_user_instances  was encountered.  See epoll(7) for further
-    /// details.
-    /// Or, The per-process limit on the number of open file descriptors has been reached.
-    ProcessFdQuotaExceeded,
-
-    /// The system-wide limit on the total number of open files has been reached.
-    SystemFdQuotaExceeded,
-
-    /// There was insufficient memory to create the kernel object.
-    SystemResources,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn linuxEpollCreate(flags: u32) LinuxEpollCreateError!i32 {
-    const rc = posix.epoll_create1(flags);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return @intCast(i32, rc),
-        else => return unexpectedErrorPosix(err),
-
-        posix.EINVAL => unreachable,
-        posix.EMFILE => return LinuxEpollCreateError.ProcessFdQuotaExceeded,
-        posix.ENFILE => return LinuxEpollCreateError.SystemFdQuotaExceeded,
-        posix.ENOMEM => return LinuxEpollCreateError.SystemResources,
-    }
-}
-
-pub const LinuxEpollCtlError = error{
-    /// op was EPOLL_CTL_ADD, and the supplied file descriptor fd is  already  registered
-    /// with this epoll instance.
-    FileDescriptorAlreadyPresentInSet,
-
-    /// fd refers to an epoll instance and this EPOLL_CTL_ADD operation would result in a
-    /// circular loop of epoll instances monitoring one another.
-    OperationCausesCircularLoop,
-
-    /// op was EPOLL_CTL_MOD or EPOLL_CTL_DEL, and fd is not registered with  this  epoll
-    /// instance.
-    FileDescriptorNotRegistered,
-
-    /// There was insufficient memory to handle the requested op control operation.
-    SystemResources,
-
-    /// The  limit  imposed  by /proc/sys/fs/epoll/max_user_watches was encountered while
-    /// trying to register (EPOLL_CTL_ADD) a new file descriptor on  an  epoll  instance.
-    /// See epoll(7) for further details.
-    UserResourceLimitReached,
-
-    /// The target file fd does not support epoll.  This error can occur if fd refers to,
-    /// for example, a regular file or a directory.
-    FileDescriptorIncompatibleWithEpoll,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn linuxEpollCtl(epfd: i32, op: u32, fd: i32, event: *linux.epoll_event) LinuxEpollCtlError!void {
-    const rc = posix.epoll_ctl(epfd, op, fd, event);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return,
-        else => return unexpectedErrorPosix(err),
-
-        posix.EBADF => unreachable, // always a race condition if this happens
-        posix.EEXIST => return LinuxEpollCtlError.FileDescriptorAlreadyPresentInSet,
-        posix.EINVAL => unreachable,
-        posix.ELOOP => return LinuxEpollCtlError.OperationCausesCircularLoop,
-        posix.ENOENT => return LinuxEpollCtlError.FileDescriptorNotRegistered,
-        posix.ENOMEM => return LinuxEpollCtlError.SystemResources,
-        posix.ENOSPC => return LinuxEpollCtlError.UserResourceLimitReached,
-        posix.EPERM => return LinuxEpollCtlError.FileDescriptorIncompatibleWithEpoll,
-    }
-}
-
-pub fn linuxEpollWait(epfd: i32, events: []linux.epoll_event, timeout: i32) usize {
-    while (true) {
-        const rc = posix.epoll_wait(epfd, events.ptr, @intCast(u32, events.len), timeout);
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0 => return rc,
-            posix.EINTR => continue,
-            posix.EBADF => unreachable,
-            posix.EFAULT => unreachable,
-            posix.EINVAL => unreachable,
-            else => unreachable,
-        }
-    }
-}
-
-pub const LinuxEventFdError = error{
-    InvalidFlagValue,
-    SystemResources,
-    ProcessFdQuotaExceeded,
-    SystemFdQuotaExceeded,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn linuxEventFd(initval: u32, flags: u32) LinuxEventFdError!i32 {
-    const rc = posix.eventfd(initval, flags);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return @intCast(i32, rc),
-        else => return unexpectedErrorPosix(err),
-
-        posix.EINVAL => return LinuxEventFdError.InvalidFlagValue,
-        posix.EMFILE => return LinuxEventFdError.ProcessFdQuotaExceeded,
-        posix.ENFILE => return LinuxEventFdError.SystemFdQuotaExceeded,
-        posix.ENODEV => return LinuxEventFdError.SystemResources,
-        posix.ENOMEM => return LinuxEventFdError.SystemResources,
-    }
-}
-
-pub const PosixGetSockNameError = error{
-    /// Insufficient resources were available in the system to perform the operation.
-    SystemResources,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn posixGetSockName(sockfd: i32) PosixGetSockNameError!posix.sockaddr {
-    var addr: posix.sockaddr = undefined;
-    var addrlen: posix.socklen_t = @sizeOf(posix.sockaddr);
-    const rc = posix.getsockname(sockfd, &addr, &addrlen);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return addr,
-        else => return unexpectedErrorPosix(err),
-
-        posix.EBADF => unreachable,
-        posix.EFAULT => unreachable,
-        posix.EINVAL => unreachable,
-        posix.ENOTSOCK => unreachable,
-        posix.ENOBUFS => return PosixGetSockNameError.SystemResources,
-    }
-}
-
-pub const PosixConnectError = error{
-    /// For UNIX domain sockets, which are identified by pathname: Write permission is denied on  the  socket
-    /// file,  or  search  permission  is  denied  for  one of the directories in the path prefix.
-    /// or
-    /// The user tried to connect to a broadcast address without having the socket broadcast flag enabled  or
-    /// the connection request failed because of a local firewall rule.
-    PermissionDenied,
-
-    /// Local address is already in use.
-    AddressInUse,
-
-    /// (Internet  domain  sockets)  The  socket  referred  to  by sockfd had not previously been bound to an
-    /// address and, upon attempting to bind it to an ephemeral port, it was determined that all port numbers
-    /// in    the    ephemeral    port    range    are   currently   in   use.    See   the   discussion   of
-    /// /proc/sys/net/ipv4/ip_local_port_range in ip(7).
-    AddressNotAvailable,
-
-    /// The passed address didn't have the correct address family in its sa_family field.
-    AddressFamilyNotSupported,
-
-    /// Insufficient entries in the routing cache.
-    SystemResources,
-
-    /// A connect() on a stream socket found no one listening on the remote address.
-    ConnectionRefused,
-
-    /// Network is unreachable.
-    NetworkUnreachable,
-
-    /// Timeout  while  attempting  connection.   The server may be too busy to accept new connections.  Note
-    /// that for IP sockets the timeout may be very long when syncookies are enabled on the server.
-    ConnectionTimedOut,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn posixConnect(sockfd: i32, sockaddr: *const posix.sockaddr) PosixConnectError!void {
-    while (true) {
-        const rc = posix.connect(sockfd, sockaddr, @sizeOf(posix.sockaddr));
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0 => return,
-            else => return unexpectedErrorPosix(err),
-
-            posix.EACCES => return PosixConnectError.PermissionDenied,
-            posix.EPERM => return PosixConnectError.PermissionDenied,
-            posix.EADDRINUSE => return PosixConnectError.AddressInUse,
-            posix.EADDRNOTAVAIL => return PosixConnectError.AddressNotAvailable,
-            posix.EAFNOSUPPORT => return PosixConnectError.AddressFamilyNotSupported,
-            posix.EAGAIN => return PosixConnectError.SystemResources,
-            posix.EALREADY => unreachable, // The socket is nonblocking and a previous connection attempt has not yet been completed.
-            posix.EBADF => unreachable, // sockfd is not a valid open file descriptor.
-            posix.ECONNREFUSED => return PosixConnectError.ConnectionRefused,
-            posix.EFAULT => unreachable, // The socket structure address is outside the user's address space.
-            posix.EINPROGRESS => unreachable, // The socket is nonblocking and the connection cannot be completed immediately.
-            posix.EINTR => continue,
-            posix.EISCONN => unreachable, // The socket is already connected.
-            posix.ENETUNREACH => return PosixConnectError.NetworkUnreachable,
-            posix.ENOTSOCK => unreachable, // The file descriptor sockfd does not refer to a socket.
-            posix.EPROTOTYPE => unreachable, // The socket type does not support the requested communications protocol.
-            posix.ETIMEDOUT => return PosixConnectError.ConnectionTimedOut,
-        }
-    }
-}
-
-/// Same as posixConnect except it is for blocking socket file descriptors.
-/// It expects to receive EINPROGRESS.
-pub fn posixConnectAsync(sockfd: i32, sockaddr: *const c_void, len: u32) PosixConnectError!void {
-    while (true) {
-        const rc = posix.connect(sockfd, sockaddr, len);
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0, posix.EINPROGRESS => return,
-            else => return unexpectedErrorPosix(err),
-
-            posix.EACCES => return PosixConnectError.PermissionDenied,
-            posix.EPERM => return PosixConnectError.PermissionDenied,
-            posix.EADDRINUSE => return PosixConnectError.AddressInUse,
-            posix.EADDRNOTAVAIL => return PosixConnectError.AddressNotAvailable,
-            posix.EAFNOSUPPORT => return PosixConnectError.AddressFamilyNotSupported,
-            posix.EAGAIN => return PosixConnectError.SystemResources,
-            posix.EALREADY => unreachable, // The socket is nonblocking and a previous connection attempt has not yet been completed.
-            posix.EBADF => unreachable, // sockfd is not a valid open file descriptor.
-            posix.ECONNREFUSED => return PosixConnectError.ConnectionRefused,
-            posix.EFAULT => unreachable, // The socket structure address is outside the user's address space.
-            posix.EINTR => continue,
-            posix.EISCONN => unreachable, // The socket is already connected.
-            posix.ENETUNREACH => return PosixConnectError.NetworkUnreachable,
-            posix.ENOTSOCK => unreachable, // The file descriptor sockfd does not refer to a socket.
-            posix.EPROTOTYPE => unreachable, // The socket type does not support the requested communications protocol.
-            posix.ETIMEDOUT => return PosixConnectError.ConnectionTimedOut,
-        }
-    }
-}
-
-pub fn posixGetSockOptConnectError(sockfd: i32) PosixConnectError!void {
-    var err_code: i32 = undefined;
-    var size: u32 = @sizeOf(i32);
-    const rc = posix.getsockopt(sockfd, posix.SOL_SOCKET, posix.SO_ERROR, @ptrCast([*]u8, &err_code), &size);
-    assert(size == 4);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => switch (err_code) {
-            0 => return,
-            else => return unexpectedErrorPosix(err),
-
-            posix.EACCES => return PosixConnectError.PermissionDenied,
-            posix.EPERM => return PosixConnectError.PermissionDenied,
-            posix.EADDRINUSE => return PosixConnectError.AddressInUse,
-            posix.EADDRNOTAVAIL => return PosixConnectError.AddressNotAvailable,
-            posix.EAFNOSUPPORT => return PosixConnectError.AddressFamilyNotSupported,
-            posix.EAGAIN => return PosixConnectError.SystemResources,
-            posix.EALREADY => unreachable, // The socket is nonblocking and a previous connection attempt has not yet been completed.
-            posix.EBADF => unreachable, // sockfd is not a valid open file descriptor.
-            posix.ECONNREFUSED => return PosixConnectError.ConnectionRefused,
-            posix.EFAULT => unreachable, // The socket structure address is outside the user's address space.
-            posix.EISCONN => unreachable, // The socket is already connected.
-            posix.ENETUNREACH => return PosixConnectError.NetworkUnreachable,
-            posix.ENOTSOCK => unreachable, // The file descriptor sockfd does not refer to a socket.
-            posix.EPROTOTYPE => unreachable, // The socket type does not support the requested communications protocol.
-            posix.ETIMEDOUT => return PosixConnectError.ConnectionTimedOut,
-        },
-        else => return unexpectedErrorPosix(err),
-        posix.EBADF => unreachable, // The argument sockfd is not a valid file descriptor.
-        posix.EFAULT => unreachable, // The address pointed to by optval or optlen is not in a valid part of the process address space.
-        posix.EINVAL => unreachable,
-        posix.ENOPROTOOPT => unreachable, // The option is unknown at the level indicated.
-        posix.ENOTSOCK => unreachable, // The file descriptor sockfd does not refer to a socket.
     }
 }
 
@@ -3123,7 +1555,6 @@ pub const SpawnThreadError = error{
     /// Not enough userland memory to spawn the thread.
     OutOfMemory,
 
-    /// See https://github.com/ziglang/zig/issues/1396
     Unexpected,
 };
 
@@ -3301,47 +1732,16 @@ pub fn spawnThread(context: var, comptime startFn: var) SpawnThreadError!*Thread
     }
 }
 
-pub fn posixWait(pid: i32) i32 {
-    var status: i32 = undefined;
-    while (true) {
-        const err = posix.getErrno(posix.waitpid(pid, &status, 0));
-        switch (err) {
-            0 => return status,
-            posix.EINTR => continue,
-            posix.ECHILD => unreachable, // The process specified does not exist. It would be a race condition to handle this error.
-            posix.EINVAL => unreachable, // The options argument was invalid
-            else => unreachable,
-        }
-    }
-}
-
-pub fn posixFStat(fd: i32) !posix.Stat {
-    var stat: posix.Stat = undefined;
-    const err = posix.getErrno(posix.fstat(fd, &stat));
-    if (err > 0) {
-        return switch (err) {
-            // We do not make this an error code because if you get EBADF it's always a bug,
-            // since the fd could have been reused.
-            posix.EBADF => unreachable,
-            posix.ENOMEM => error.SystemResources,
-            else => os.unexpectedErrorPosix(err),
-        };
-    }
-
-    return stat;
-}
-
 pub const CpuCountError = error{
     OutOfMemory,
     PermissionDenied,
 
-    /// See https://github.com/ziglang/zig/issues/1396
     Unexpected,
 };
 
 pub fn cpuCount(fallback_allocator: *mem.Allocator) CpuCountError!usize {
     switch (builtin.os) {
-        builtin.Os.macosx, builtin.Os.freebsd, builtin.Os.netbsd => {
+        .macosx, .freebsd, .netbsd => {
             var count: c_int = undefined;
             var count_len: usize = @sizeOf(c_int);
             const rc = posix.sysctlbyname(switch (builtin.os) {
@@ -3361,7 +1761,7 @@ pub fn cpuCount(fallback_allocator: *mem.Allocator) CpuCountError!usize {
                 else => return os.unexpectedErrorPosix(err),
             }
         },
-        builtin.Os.linux => {
+        .linux => {
             const usize_count = 16;
             const allocator = std.heap.stackFallback(usize_count * @sizeOf(usize), fallback_allocator).get();
 
@@ -3393,136 +1793,11 @@ pub fn cpuCount(fallback_allocator: *mem.Allocator) CpuCountError!usize {
                 }
             }
         },
-        builtin.Os.windows => {
+        .windows => {
             var system_info: windows.SYSTEM_INFO = undefined;
             windows.GetSystemInfo(&system_info);
             return @intCast(usize, system_info.dwNumberOfProcessors);
         },
         else => @compileError("unsupported OS"),
-    }
-}
-
-pub const BsdKQueueError = error{
-    /// The per-process limit on the number of open file descriptors has been reached.
-    ProcessFdQuotaExceeded,
-
-    /// The system-wide limit on the total number of open files has been reached.
-    SystemFdQuotaExceeded,
-
-    /// See https://github.com/ziglang/zig/issues/1396
-    Unexpected,
-};
-
-pub fn bsdKQueue() BsdKQueueError!i32 {
-    const rc = posix.kqueue();
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return @intCast(i32, rc),
-        posix.EMFILE => return BsdKQueueError.ProcessFdQuotaExceeded,
-        posix.ENFILE => return BsdKQueueError.SystemFdQuotaExceeded,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub const BsdKEventError = error{
-    /// The process does not have permission to register a filter.
-    AccessDenied,
-
-    /// The event could not be found to be modified or deleted.
-    EventNotFound,
-
-    /// No memory was available to register the event.
-    SystemResources,
-
-    /// The specified process to attach to does not exist.
-    ProcessNotFound,
-};
-
-pub fn bsdKEvent(
-    kq: i32,
-    changelist: []const posix.Kevent,
-    eventlist: []posix.Kevent,
-    timeout: ?*const posix.timespec,
-) BsdKEventError!usize {
-    while (true) {
-        const rc = posix.kevent(kq, changelist, eventlist, timeout);
-        const err = posix.getErrno(rc);
-        switch (err) {
-            0 => return rc,
-            posix.EACCES => return BsdKEventError.AccessDenied,
-            posix.EFAULT => unreachable,
-            posix.EBADF => unreachable,
-            posix.EINTR => continue,
-            posix.EINVAL => unreachable,
-            posix.ENOENT => return BsdKEventError.EventNotFound,
-            posix.ENOMEM => return BsdKEventError.SystemResources,
-            posix.ESRCH => return BsdKEventError.ProcessNotFound,
-            else => unreachable,
-        }
-    }
-}
-
-pub fn linuxINotifyInit1(flags: u32) !i32 {
-    const rc = linux.inotify_init1(flags);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return @intCast(i32, rc),
-        posix.EINVAL => unreachable,
-        posix.EMFILE => return error.ProcessFdQuotaExceeded,
-        posix.ENFILE => return error.SystemFdQuotaExceeded,
-        posix.ENOMEM => return error.SystemResources,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub fn linuxINotifyAddWatchC(inotify_fd: i32, pathname: [*]const u8, mask: u32) !i32 {
-    const rc = linux.inotify_add_watch(inotify_fd, pathname, mask);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return @intCast(i32, rc),
-        posix.EACCES => return error.AccessDenied,
-        posix.EBADF => unreachable,
-        posix.EFAULT => unreachable,
-        posix.EINVAL => unreachable,
-        posix.ENAMETOOLONG => return error.NameTooLong,
-        posix.ENOENT => return error.FileNotFound,
-        posix.ENOMEM => return error.SystemResources,
-        posix.ENOSPC => return error.UserResourceLimitReached,
-        else => return unexpectedErrorPosix(err),
-    }
-}
-
-pub fn linuxINotifyRmWatch(inotify_fd: i32, wd: i32) !void {
-    const rc = linux.inotify_rm_watch(inotify_fd, wd);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return rc,
-        posix.EBADF => unreachable,
-        posix.EINVAL => unreachable,
-        else => unreachable,
-    }
-}
-
-pub const MProtectError = error{
-    AccessDenied,
-    OutOfMemory,
-    Unexpected,
-};
-
-/// address and length must be page-aligned
-pub fn posixMProtect(address: usize, length: usize, protection: u32) MProtectError!void {
-    const negative_page_size = @bitCast(usize, -isize(page_size));
-    const aligned_address = address & negative_page_size;
-    const aligned_end = (address + length + page_size - 1) & negative_page_size;
-    assert(address == aligned_address);
-    assert(length == aligned_end - aligned_address);
-    const rc = posix.mprotect(address, length, protection);
-    const err = posix.getErrno(rc);
-    switch (err) {
-        0 => return,
-        posix.EINVAL => unreachable,
-        posix.EACCES => return error.AccessDenied,
-        posix.ENOMEM => return error.OutOfMemory,
-        else => return unexpectedErrorPosix(err),
     }
 }
