+++ std/os/windows/errno.zig
@@ -1,0 +1,1 @@
+// TODO get these values from msvcrt
